+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Potential crash when parsing js files in non-ultimate versions.
 - Better formatting of javascript files in non-ultimate versions.
+- Better formatting of javascript files in ultimate.
 
 ### From 2.0.0